testutils: use generated code
Instead of the previous hand-written code. This allows us to
automatically adapt to new event types.
